third_party/blink/web_tests/platform/mac-mac11-arm64/virtual/expose-lcp-render-time/external/wpt/performance-timeline/cross-origin-non-tao-image.sub-expected.txt

This is a testharness.js-based test.
FAIL lcp and fcp assert_approx_equals: expected 70.79999999981374 +/- 0.001 but got 72.399
Harness: the test ran to completion.
